Move LoadingSpinner to SEE.UI
