--- Assets/SEE/UI/LoadingSpinner.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 24947e2c27f84845bb5b50f4da903984
-timeCreated: 1702573624
